# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "express-api",
    "version": "1.0.0",
    "description": "express-api"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "unauthorizedError": {
        "description": "Access token is missing or invalid"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "example": 1 },
          "username": { "type": "string", "example": "administrator" },
          "password": { "type": "string" },
          "fullName": { "type": "string", "example": "Do Thanh Tung" },
          "phone": { "type": "string", "example": 987654321 },
          "email": { "type": "string", "example": "admin@admin.com" },
          "address": { "type": "string", "example": "Tp HCM" },
          "statusId": {
            "type": "integer",
            "description": "User Status",
            "format": "int32",
            "example": 1
          },
          "deleteFlag": { "type": "integer", "format": "int64", "example": 1 },
          "createdBy": { "type": "integer", "format": "int64", "example": 0 },
          "updatedBy": { "type": "integer", "format": "int64", "example": 0 }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/authentications/login": {
      "post": {
        "tags": ["Authentications"],
        "description": "Login to system",
        "requestBody": {
          "description": "Login with username and password",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "administrator" },
                  "password": { "type": "string", "example": "admin@123456" }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": { "description": "Invalid ID supplied" },
          "404": { "description": "Pet not found" },
          "405": { "description": "Validation exception" }
        }
      }
    },
    "/authentications/refresh-token": {
      "post": {
        "tags": ["Authentications"],
        "description": "Refresh Token",
        "requestBody": {
          "description": "Refresh Token",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "refreshToken": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "404": { "description": "User not found" },
          "405": { "description": "Validation exception" }
        }
      }
    },
    "/authentications/logout": {
      "post": {
        "tags": ["Authentications"],
        "description": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "refreshToken": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "401": { "$ref": "#/components/responses/unauthorizedError" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/users": {
      "get": {
        "tags": ["Users"],
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "post": {
        "tags": ["Users"],
        "description": "Add user" ,
        "requestBody":{
            "content": {
                "application/json":{
                    "schema":{
                        "type": "object",
                        "$ref": "#/components/schemas/User"
                    }
                }
            }
        },
        "responses": {
            "200":{
                "description": "Success",
                "content": {
                    "application/json":{
                        "schema":{
                            "type": "object",
                            "$ref": "#/components/schemas/User"
                        }
                    }
                }
            }
        },
        "security": [
            {"bearerAuth": []}
        ]
    }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "404": { "description": "Not found" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    }
  },
  "tags": [
    { "name": "Authentications", "description": "Authentication" },
    { "name": "Users", "description": "Users" }
  ]
}